Added loading image view to make it reusable
